Remove reward.addressBookConnector
